requirements.txt
alembic==1.4.2
atomicwrites==1.4.0
attrs==19.3.0
cffi==1.14.0
click==7.1.1
colorama==0.4.3
cryptography==2.9.2
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.2
MarkupSafe==1.1.1
marshmallow==3.5.2
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
py==1.8.2
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.9.3
pyparsing==2.4.7
pytest==5.4.3
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
six==1.14.0
SQLAlchemy==1.3.16
wcwidth==0.2.4
webargs==6.1.0
Werkzeug==1.0.1
